refactor(server): split effective-entity resolver one-export-per-file; trim comments to timeless notes
