+++ packages/twenty-server/src/engine/metadata-modules/field-metadata/types/field-metadata-overrides.type.ts
@@ -1,8 +1,5 @@
 import { type APP_LOCALES } from 'twenty-shared/translations';
 
-// Anonymous single-slot override blob for field metadata: flat presentation
-// overrides plus an optional per-locale translations sub-map for translatable
-// properties.
 export type FieldMetadataOverrides = {
   label?: string | null;
   description?: string | null;
