+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/types/object-metadata-overrides.type.ts
@@ -1,8 +1,5 @@
 import { type APP_LOCALES } from 'twenty-shared/translations';
 
-// Anonymous single-slot override blob for object metadata: flat presentation
-// overrides plus an optional per-locale translations sub-map for translatable
-// properties.
 export type ObjectMetadataOverrides = {
   labelSingular?: string | null;
   labelPlural?: string | null;
