+++ packages/twenty-server/src/engine/metadata-modules/utils/__tests__/resolve-effective-entity.util.spec.ts
@@ -4,10 +4,8 @@ import { type APP_LOCALES, SOURCE_LOCALE } from 'twenty-shared/translations';
 import { isDefined } from 'twenty-shared/utils';
 
 import { translateStandardLabel } from 'src/engine/core-modules/i18n/utils/translate-standard-label.util';
-import {
-  resolveEffectiveEntity,
-  resolveEffectiveEntityProperty,
-} from 'src/engine/metadata-modules/utils/resolve-effective-entity.util';
+import { resolveEffectiveEntityProperty } from 'src/engine/metadata-modules/utils/resolve-effective-entity-property.util';
+import { resolveEffectiveEntity } from 'src/engine/metadata-modules/utils/resolve-effective-entity.util';
 
 // Frozen reference implementations of the three resolvers that existed before
 // unification (resolve-object-metadata-standard-override,
