+++ packages/twenty-server/src/engine/metadata-modules/utils/effective-entity-i18n-context.type.ts
@@ -1,0 +1,9 @@
+import { type I18n } from '@lingui/core';
+import { type APP_LOCALES } from 'twenty-shared/translations';
+
+export type EffectiveEntityI18nContext = {
+  locale: keyof typeof APP_LOCALES | undefined;
+  i18nInstance: I18n;
+  isStandardApp: boolean;
+  applicationCatalog?: Record<string, string>;
+};
